+++ include/webserver/sql/MySQLConnector.hpp
@@ -7,7 +7,7 @@
 #include <cppconn/statement.h>
 #include <cppconn/resultset.h>
 #include <iostream>
-#include <../log/log.hpp>
+#include "webserver/utils/log.hpp"
 
 class MySQLConnector {
 public:
