+++ src/utils/log.cpp
@@ -1,4 +1,4 @@
-#include "log.hpp"
+#include "webserver/utils/log.hpp"
 
 Logger::Logger() : running_(false), async_(true) {}
